Adding open source notes.
