+++ src/DurableTask.CosmosDB/Documentation/orchestrationservice_dependencies.png
