+++ src/DurableTask.CosmosDB/Monitoring/AzureStorageOrchestrationServiceStats.cs
@@ -11,7 +11,7 @@
 //  limitations under the License.
 //  ----------------------------------------------------------------------------------
 
-namespace DurableTask.AzureStorage.Monitoring
+namespace DurableTask.CosmosDB.Monitoring
 {
     using DurableTask.Core.Stats;
